+++ Assets/Script/Object/Plant.cs
@@ -13,10 +13,10 @@ public class Plant : MonoBehaviour
     InteractionMenu interactionMenu;
     Monologue monologue;
     GameObject Player;
+    EnergyGauge energyGauge;
+    int sceneNum;
 
     Animator animaitor;
-
-    int sceneNum;
 
     public bool isGatherPossible;
     public string plantName = "StickPlant";
@@ -37,6 +37,7 @@ public class Plant : MonoBehaviour
         interactionMenu = GameObject.Find("InteractionMenu").GetComponent<InteractionMenu>();
         Player = GameObject.Find("Player");
         monologue = Player.transform.Find("Monologue").gameObject.GetComponent<Monologue>();
+        energyGauge = GameObject.Find("LeftUI").GetComponent<EnergyGauge>();
 
         animaitor = GetComponent<Animator>();
 
@@ -54,6 +55,7 @@ public class Plant : MonoBehaviour
 
     public void GetItem()
     {
+        energyGauge.SetAmount(-10);
         if (isTumor == false)
         {
             switch (plantName)
@@ -103,8 +105,43 @@ public class Plant : MonoBehaviour
 
     public void RemoveObject()
     {
+        if (energyGauge.GetAmount() < 10)
+        {
+            monologue.DisplayLog("에너지가 부족해서 제거할 수 없어.");
+            return;
+        }
+        energyGauge.SetAmount(-10);
         interactionIcon.DeleteAllIcons();
         SceneObjectManager.instance.DeleteObject(sceneNum, Grid.instance.PosToGrid(transform.position.x));
+    }
+
+    void Examine()
+    {
+        switch(plantName)
+        {
+            case "StickPlant":
+                monologue.DisplayLog("집게발 모양의 대나무이다.\n다행히 공격적이지는 않아서 채집하기 쉬울 것 같다.");
+                break;
+            case "MassPlant":
+                monologue.DisplayLog("괴물의 조직? 아니면 배설물?\n정체는 알 수 없지만 어딘가에 쓸 수는 있을 것 같다.");
+                break;
+            case "BoardPlant":
+                monologue.DisplayLog("이 식물은 재생속도가 비정상적으로 빠르다.\n애초에 더이상 지구에 정상적인 것은 존재하지 않는다.");
+                break;
+            case "ThornPlant":
+                monologue.DisplayLog("선인장의 변종으로 예상했지만 아니었다.\n이 식물의 가시 성분은 식물조직보다는 동물의 뿔에 가깝다.");
+                break;
+            case "Trap01":
+                if(state == 1)
+                {
+                    monologue.DisplayLog("덫에 무언가가 잡힌 것 같다!\n아이템 획득을 눌러서 확인해보자.");
+                }
+                else
+                {
+                    monologue.DisplayLog("괴식물의 가시로 만든 덫이다.\n괴식물의 둥지 근처에 설치하면 무언가 잡힐지도 모른다.");
+                }
+                break;
+        }
     }
 
     public bool InventoryCheck()
@@ -294,6 +331,7 @@ public class Plant : MonoBehaviour
                 }
             }
         }
+        interactionMenu.AddMenu(InteractionMenu.MenuItem.Examine);
         interactionMenu.AddMenu(InteractionMenu.MenuItem.Remove);
 
         float w = GetComponent<SpriteRenderer>().sprite.rect.width;
@@ -308,10 +346,22 @@ public class Plant : MonoBehaviour
             case InteractionMenu.MenuItem.Gather:
                 if (isGatherPossible == false)
                 {
-                    monologue.DisplayLog("아직 채집할 수 없어.\n조금 더 자란 다음에 채집하는게 좋겠군.");
+                    if (plantName == "Trap01")
+                    {
+                        monologue.DisplayLog("아직 덫에 아무 것도 잡히지 않았다.");
+                    }
+                    else
+                    {
+                        monologue.DisplayLog("아직 채집할 수 없어.\n조금 더 자란 다음에 채집하는게 좋겠군.");
+                    }
                 }
                 else if (InventoryCheck() == true)
                 {
+                    if (energyGauge.GetAmount() < 10)
+                    {
+                        monologue.DisplayLog("에너지가 부족해서 채집할 수 없어.");
+                        return;
+                    }
                     GatherStart();
                     Player.GetComponent<PlayerInteraction>().GatherPlant(GatherAnimationType);
                 }
@@ -325,6 +375,9 @@ public class Plant : MonoBehaviour
                 break;
             case InteractionMenu.MenuItem.Remove:
                 RemoveObject();
+                break;
+            case InteractionMenu.MenuItem.Examine:
+                Examine();
                 break;
         }
     }
